[core] Move RuleSets/RuleSetReference into internal sub-package
