+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/AnalysisCacheListener.java
@@ -11,7 +11,7 @@ import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.document.TextFile;
 import net.sourceforge.pmd.reporting.FileAnalysisListener;
 import net.sourceforge.pmd.reporting.GlobalAnalysisListener;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 
 /**
  * Adapter to wrap {@link AnalysisCache} behaviour in a {@link GlobalAnalysisListener}.
