+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/CachedRuleMapper.java
@@ -9,7 +9,7 @@ import java.util.Map;
 
 import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.rule.Rule;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 
 /**
  * A mapper from rule class names to rule instances for cached rules.
