+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/NoopAnalysisCache.java
@@ -13,7 +13,7 @@ import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.document.TextDocument;
 import net.sourceforge.pmd.lang.document.TextFile;
 import net.sourceforge.pmd.reporting.FileAnalysisListener;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 
 /**
  * A NOOP analysis cache. Easier / safer than null-checking.
