+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/LanguageProcessor.java
@@ -13,7 +13,7 @@ import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.cache.AnalysisCache;
 import net.sourceforge.pmd.lang.document.TextFile;
 import net.sourceforge.pmd.reporting.GlobalAnalysisListener;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 import net.sourceforge.pmd.util.log.MessageReporter;
 
 /**
