+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/impl/MonoThreadProcessor.java
@@ -6,7 +6,7 @@ package net.sourceforge.pmd.lang.impl;
 
 import net.sourceforge.pmd.lang.LanguageProcessor.AnalysisTask;
 import net.sourceforge.pmd.lang.document.TextFile;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 
 /**
  * @author Romain Pelisse &lt;belaran@gmail.com&gt;
