+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/RuleReference.java
@@ -16,7 +16,7 @@ import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.rule.Rule;
 import net.sourceforge.pmd.rule.RulePriority;
-import net.sourceforge.pmd.rule.RuleSetReference;
+import net.sourceforge.pmd.rule.internal.RuleSetReference;
 import net.sourceforge.pmd.util.StringUtil;
 
 /**
