+++ pmd-core/src/main/java/net/sourceforge/pmd/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+/**
+ * This is the PMD programming mistake detector.
+ *
+ * <p>Use it through {@link net.sourceforge.pmd.PmdAnalysis} and
+ * {@link net.sourceforge.pmd.cpd.CpdAnalysis}.
+ */
+package net.sourceforge.pmd;
