--- pmd-core/src/main/java/net/sourceforge/pmd/package.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html "-//W3C//DTD HTML 3.2 Final//EN">
-<html>
-    <head>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <title>PMD</title>
-    </head>
-    <body>
-        <p>This is the PMD programming mistake detector.  In order to use it,
-            create a new instance of {@link net.sourceforge.pmd.PMD} and use its
-            {@link net.sourceforge.pmd.PMD#processFiles} method.</p>
-    </body>
-</html>
