+++ pmd-core/src/main/java/net/sourceforge/pmd/rule/internal/RuleSetReference.java
@@ -2,21 +2,19 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.rule;
+package net.sourceforge.pmd.rule.internal;
 
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-import net.sourceforge.pmd.annotation.InternalApi;
+import net.sourceforge.pmd.rule.RuleSetLoader;
 
 /**
  * This class represents a reference to RuleSet.
  *
- * @deprecated This is part of the internals of the {@link RuleSetLoader}.
+ * <p>Note: This is part of the internals of the {@link RuleSetLoader}.
  */
-@Deprecated
-@InternalApi
 public class RuleSetReference {
     private final String ruleSetFileName;
     private final boolean allRules;
