+++ pmd-core/src/main/java/net/sourceforge/pmd/rule/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+/**
+ * Contains {@link net.sourceforge.pmd.rule.Rule} and {@link net.sourceforge.pmd.rule.RuleSet}
+ * related classes.
+ */
+package net.sourceforge.pmd.rule;
