+++ pmd-core/src/test/java/net/sourceforge/pmd/ReportTestUtil.java
@@ -12,7 +12,7 @@ import net.sourceforge.pmd.lang.ast.RootNode;
 import net.sourceforge.pmd.lang.document.TestMessageReporter;
 import net.sourceforge.pmd.rule.Rule;
 import net.sourceforge.pmd.rule.RuleSet;
-import net.sourceforge.pmd.rule.RuleSets;
+import net.sourceforge.pmd.rule.internal.RuleSets;
 
 public final class ReportTestUtil {
     private ReportTestUtil() {
