+++ pmd-core/src/test/java/net/sourceforge/pmd/rule/RuleReferenceTest.java
@@ -19,6 +19,7 @@ import net.sourceforge.pmd.lang.rule.MockRule;
 import net.sourceforge.pmd.lang.rule.RuleReference;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.properties.PropertyFactory;
+import net.sourceforge.pmd.rule.internal.RuleSetReference;
 
 class RuleReferenceTest {
